# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inventory Management System API",
    "version": "1.0.0",
    "description": "Simple products & stock API for ASE challenge"
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health check",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/products": {
      "get": {
        "summary": "List all products",
        "responses": {
          "200": { "description": "OK" }
        }
      },
      "post": {
        "summary": "Create a product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "stock_quantity": { "type": "integer", "minimum": 0 },
                  "low_stock_threshold": { "type": "integer", "minimum": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Get one product",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update a product",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "stock_quantity": { "type": "integer", "minimum": 0 },
                  "low_stock_threshold": { "type": "integer", "minimum": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Validation error" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete a product",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/products/{id}/increase-stock": {
      "post": {
        "summary": "Increase stock",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["amount"], "properties": { "amount": { "type": "integer", "minimum": 1 } } }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Validation error" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/products/{id}/decrease-stock": {
      "post": {
        "summary": "Decrease stock",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["amount"], "properties": { "amount": { "type": "integer", "minimum": 1 } } }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Validation error / insufficient stock" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/products/low/threshold": {
      "get": {
        "summary": "List products below threshold",
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}